Add missing author name to copyright info
Dave Rolsky had added his name to most copyright statements in an
earlier commit, however had added his name to the LICENSE file, which is
actually automatically generated.  Hence his name needed to be added to
`dist.ini`.  The LICENSE is also updated in this commit since it is also
part of the repository.  Instead of simply replacing the current year as
is the default in Dist::Zilla the initial year (found to be 2014) is
added explicitly.  Also, since there hasn't yet been a release of this
software in 2016, the final year in the copyright statement has been
left at the most recent release year, 2015.